`get-rfc-grep-rfc-index' now deals with rfc-index.txt not being available locally.
New function -- `get-rfc-grep-rfc-index'.

New variable -- `get-rfc-open-in-new-frame'.  Make opening an RFC in a
new frame a user option.

Added customize support.

New variable -- `get-rfc-view-rfc-mode'.  If this is set, it is the
name of the mode we want to view RFC's in.
Modified `get-rfc-view-rfc' slightly to make use of this.
New variable -- `get-rfc-no-wget'.  If this is non-nil, we don't
try and fetch a remote RFC via wget, but rather call `browse-url'.

New function -- `get-rfc'.  This allows us to get an rfc from
a remote site and then display it in Emacs.