Rework all functions to take a time-spec.
All functions that deal with time values now take a value like that
returned from decode-time, rather than a mismash of time-specs and
explicit month and year arguments.

Additionally, we no longer special case a null argument to
timex-parse-files to mean the current month of files, the list of
files must be specified explicitly.