New function timex-increment-time.
This function takes care of wrapping around when a particular value
overflows, for example incrementing the day in a time-spec past the
last day of the month should move to the next month.

Also fix boundary case in timex-split-date-range.  If d2 was the last
day of the month, seconds and so on became important in deciding if to
split a date range.  This was the wrong thing to do.  To fix, copy
unimportant bits (seconds, minutes and hours) from d2 into d1.